+++ R/auk-package.r
@@ -1,4 +1,4 @@
-#' `auk`: eBird Data Extraction and Processing with AWK
+#' `auk`: eBird Data Extraction and Processing in R
 #'
 #' Tools for extracting and processing eBird data from the eBird Basic Dataset 
 #' (EBD).