+++ R/auk-rollup.r
@@ -12,9 +12,9 @@
 #'
 #' @param x data.frame; data frame of eBird data, typically as imported by
 #'   [read_ebd()]
-#' @param version integer; the version (i.e. year) of the taxonomy. Leave empty 
-#'   to use the version of the taxonomy included in the packages. See 
-#'   [get_ebird_taxonomy()]. 
+#' @param taxonomy_version integer; the version (i.e. year) of the taxonomy. In
+#'   most cases, this should be left empty to use the version of the taxonomy
+#'   included in the package. See [get_ebird_taxonomy()].
 #'   
 #' @details When rolling observations up to species level the observed counts
 #'   are summed across any taxa that resolve to the same species. However, if
@@ -48,6 +48,13 @@
 #'   free-flying (these do not count on personal lists) e.g., Mallard (Domestic
 #'   type). Dropped by `auk_rollup()`.
 #'   
+#'   The rollup process is based on the eBird taxonomy, which is updated once a
+#'   year in August. The `auk` package includes a copy of the eBird taxonomy,
+#'   current at the time of release; however, if the EBD and `auk` versions are
+#'   not aligned, you may need to explicitly specify which version of the
+#'   taxonomy to use, in which case the eBird API will be queried to get the
+#'   correct version of the taxonomy.
+#'   
 #' @return A data frame of the eBird data with taxonomic rollup applied.
 #' @references Consult the [eBird taxonomy](http://help.ebird.org/customer/portal/articles/1006825-the-ebird-taxonomy) 
 #'   page for further details.
@@ -77,7 +84,7 @@
 #' ebd_ru %>%
 #'   filter(common_name == "Yellow-rumped Warbler") %>% 
 #'   select(checklist_id, category, common_name, observation_count)
-auk_rollup <- function(x, version) {
+auk_rollup <- function(x, taxonomy_version) {
   assertthat::assert_that(
     is.data.frame(x),
     "scientific_name" %in% names(x)
@@ -96,13 +103,12 @@ auk_rollup <- function(x, version) {
   }
   
   # get the correct ebird taxonomy version
-  if (missing(version) || version == auk_version()$taxonomy_version) {
+  if (missing(taxonomy_version) || 
+      taxonomy_version == auk_version()$taxonomy_version) {
     tax <- auk::ebird_taxonomy
   } else {
-    assertthat::assert_that(
-      is_integer(version), 
-      length(version) == 1)
-    tax <- get_ebird_taxonomy(version = version)
+    stopifnot(is_integer(taxonomy_version), length(taxonomy_version) == 1)
+    tax <- get_ebird_taxonomy(version = taxonomy_version)
   }
   
   # remove anything not identifiable to a species